create and prepare new `SimulatorsAndWallets` classes (#16931)
