+++ tests/core/mempool/test_mempool_performance.py
@@ -11,7 +11,7 @@ from chia.util.ints import uint32, uint64, uint128
 from chia.wallet.util.tx_config import DEFAULT_TX_CONFIG
 from chia.wallet.wallet_node import WalletNode
 from tests.util.misc import BenchmarkRunner
-from tests.util.setup_nodes import SimulatorsAndWallets
+from tests.util.setup_nodes import OldSimulatorsAndWallets
 from tests.util.time_out_assert import time_out_assert
 
 
@@ -28,7 +28,7 @@ async def wallet_balance_at_least(wallet_node: WalletNode, balance: uint128) -> 
 @pytest.mark.limit_consensus_modes(reason="benchmark")
 @pytest.mark.anyio
 async def test_mempool_update_performance(
-    wallet_nodes_mempool_perf: SimulatorsAndWallets,
+    wallet_nodes_mempool_perf: OldSimulatorsAndWallets,
     default_400_blocks: List[FullBlock],
     self_hostname: str,
     benchmark_runner: BenchmarkRunner,
