+++ tests/wallet/cat_wallet/test_trades.py
@@ -26,8 +26,8 @@ from chia.wallet.vc_wallet.cr_cat_drivers import ProofsChecker
 from chia.wallet.vc_wallet.cr_cat_wallet import CRCATWallet
 from chia.wallet.vc_wallet.vc_store import VCProofs
 from tests.conftest import SOFTFORK_HEIGHTS, ConsensusMode
+from tests.environments.wallet import WalletEnvironment, WalletStateTransition, WalletTestFramework
 from tests.util.time_out_assert import time_out_assert
-from tests.wallet.conftest import WalletEnvironment, WalletStateTransition, WalletTestFramework
 from tests.wallet.vc_wallet.test_vc_wallet import mint_cr_cat
 
 
@@ -91,8 +91,8 @@ async def test_cat_trades(
     # Setup
     env_maker: WalletEnvironment = wallet_environments.environments[0]
     env_taker: WalletEnvironment = wallet_environments.environments[1]
-    wallet_node_maker = env_maker.wallet_node
-    wallet_node_taker = env_taker.wallet_node
+    wallet_node_maker = env_maker.node
+    wallet_node_taker = env_taker.node
     client_maker = env_maker.rpc_client
     client_taker = env_taker.rpc_client
     wallet_maker = env_maker.xch_wallet
