+++ tests/wallet/db_wallet/test_dl_wallet.py
@@ -19,7 +19,7 @@ from chia.wallet.db_wallet.db_wallet_puzzles import create_mirror_puzzle
 from chia.wallet.util.merkle_tree import MerkleTree
 from chia.wallet.util.tx_config import DEFAULT_TX_CONFIG
 from tests.conftest import ConsensusMode
-from tests.util.setup_nodes import SimulatorsAndWallets
+from tests.util.setup_nodes import OldSimulatorsAndWallets
 from tests.util.time_out_assert import time_out_assert
 
 pytestmark = pytest.mark.data_layer
@@ -46,7 +46,7 @@ class TestDLWallet:
     )
     @pytest.mark.anyio
     async def test_initial_creation(
-        self, self_hostname: str, simulator_and_wallet: SimulatorsAndWallets, trusted: bool, reuse_puzhash: bool
+        self, self_hostname: str, simulator_and_wallet: OldSimulatorsAndWallets, trusted: bool, reuse_puzhash: bool
     ) -> None:
         full_nodes, wallets, _ = simulator_and_wallet
         full_node_api = full_nodes[0]
@@ -96,7 +96,7 @@ class TestDLWallet:
     )
     @pytest.mark.anyio
     async def test_get_owned_singletons(
-        self, self_hostname: str, simulator_and_wallet: SimulatorsAndWallets, trusted: bool
+        self, self_hostname: str, simulator_and_wallet: OldSimulatorsAndWallets, trusted: bool
     ) -> None:
         full_nodes, wallets, _ = simulator_and_wallet
         full_node_api = full_nodes[0]
@@ -148,7 +148,7 @@ class TestDLWallet:
     @pytest.mark.parametrize("trusted", [True, False])
     @pytest.mark.anyio
     async def test_tracking_non_owned(
-        self, self_hostname: str, two_wallet_nodes: SimulatorsAndWallets, trusted: bool
+        self, self_hostname: str, two_wallet_nodes: OldSimulatorsAndWallets, trusted: bool
     ) -> None:
         full_nodes, wallets, _ = two_wallet_nodes
         full_node_api = full_nodes[0]
@@ -233,7 +233,7 @@ class TestDLWallet:
     )
     @pytest.mark.anyio
     async def test_lifecycle(
-        self, self_hostname: str, simulator_and_wallet: SimulatorsAndWallets, trusted: bool
+        self, self_hostname: str, simulator_and_wallet: OldSimulatorsAndWallets, trusted: bool
     ) -> None:
         full_nodes, wallets, _ = simulator_and_wallet
         full_node_api = full_nodes[0]
@@ -338,7 +338,7 @@ class TestDLWallet:
     async def test_rebase(
         self,
         self_hostname: str,
-        two_wallet_nodes: SimulatorsAndWallets,
+        two_wallet_nodes: OldSimulatorsAndWallets,
         trusted: bool,
     ) -> None:  # pragma: no cover
         full_nodes, wallets, _ = two_wallet_nodes
